feat: 모임상세> 좋아요 구

+++ src/main/java/com/kosta/readdam/repository/ClassLikeRepository.java
@@ -15,5 +15,5 @@ public interface ClassLikeRepository extends JpaRepository<ClassLike, Integer> {
 
 	List<ClassLike> findAllByUserUsername(String username);
 
-	long countByClassIdClassId(Integer classId);
+	Integer countByClassIdClassId(Integer classId);
 }
